Refactor to reduce duplicate code for collecting hrsl and gee datasets
