+++ src/data_processing/admin_bounds.py
@@ -1,0 +1,22 @@
+import geopandas as gpd
+
+
+def join_admin_bounds(base_df, admin_bounds_gdf):
+    base_gdf = gpd.GeoDataFrame(
+        base_df,
+        geometry=gpd.points_from_xy(base_df["longitude"], base_df["latitude"]),
+        crs="EPSG:4326",
+    )
+    # First, we want to retain only the unique columns in the right GDF to avoid duplicate columns (and consequent column renames).
+    cols_unique_to_adm_bounds = admin_bounds_gdf.columns.difference(
+        base_gdf.columns
+    ).tolist() + ["geometry"]
+    admin_bounds_gdf = admin_bounds_gdf[cols_unique_to_adm_bounds]
+
+    # Perform the actual spatial join
+    base_gdf = gpd.sjoin(base_gdf, admin_bounds_gdf, predicate="within", how="left")
+
+    # Drop unnecessary columns generated by the sjoin
+    base_gdf.drop(["index_right"], axis=1, inplace=True)
+
+    return base_gdf
